pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sigen"
version = "4.0.0"
description = "Async Python client for the Sigenergy cloud API"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9.0",
    "pycryptodome>=3.19.0",
    "aiomqtt>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
